Frobenius fixes finished. I had to make some rather drastic notation changes.
I started by making the meaning of \Frob precise and consistent with
Deligne's definition of Weil sheaf. Deligne (and everyone else I could
find, like Laumon, Lusztig, Keihl and Weissauer, to name a few) uses F,
not Fr, for what we meant. So I had to change Fr to F in our paper. But
that made our current use of F, in the datum of a character sheaf, a
notational conflict. So I followed Lusztig in replacing F with \phi. I
also decided to harmonize our notation for trace of Frobenius with
Laumon, replacing our current t^Fr_L with t_L. That made a conflict
with our use of t_L in the proposition on etale group schemes, so I
changes t_L there to \tau_L, which works better in the context of
cocycles. Other than that, only minor changes ;-)