Working on \S 3.
I’m in the middle of working on \S 3, but running out of steam so I’ll
have to put it down for now. The current version is pretty butchered.
Sorry. I’ve cut much of material on finite quasicharacter sheaves and
pasted in into the file below \end{document} for later use.