Change to Introduction and Acknowledgements
In the Introduction, I moved the internal reference and the tag to the
section on quasicharacters in an attempt to match your style, and also
inserted the word commutative, where is was needed (line 317).
In the Acknowledgements, I replaced the reference to BIRS with Banff
International Research Station, since some readers may not know what
BIRS stands for.